test(model): split model-types by behaviour family under the 400-LOC ceiling
Audit per spec 0640 design Decision 6: model-types.test.js (448 LOC) is
broad per-helper coverage, not a single cross-multiplied matrix. The
*MeetsRequirement comparisons already carry only representative cases, so
there is nothing to collapse. Disposition is the Part 03 large-file split
rule: split by behaviour family into Type Helpers (model-types.test.js,
index/clamp/meets-requirement/compare/sort/group), Capability Functions +
emoji (model-types-capability.test.js), and derive* (model-types-
derivation.test.js). All 33 cases preserved verbatim; each file <=400 LOC
(192/127/134). No covered code path lost.

Refs spec 0640 (test-side hygiene) Success Criteria 4 and 5.
