+++ tests/model-types.test.js
@@ -9,31 +9,12 @@ import {
   skillProficiencyMeetsRequirement,
   behaviourMaturityMeetsRequirement,
   groupSkillsByCapability,
-  getCapabilityById,
-  getCapabilityOrder,
-  getCapabilityEmoji,
-  getCapabilityResponsibility,
-  getConceptEmoji,
 } from "@forwardimpact/map/levels";
 
 import {
   compareByCapability,
   sortSkillsByCapability,
 } from "@forwardimpact/libskill/policies";
-
-import {
-  deriveResponsibilities,
-  deriveJob,
-} from "@forwardimpact/libskill/derivation";
-
-import {
-  testSkills,
-  testBehaviours,
-  testDiscipline,
-  testTrack,
-  testLevel,
-  testCategories,
-} from "./model-fixtures.js";
 
 describe("Type Helpers", () => {
   describe("getSkillProficiencyIndex", () => {
@@ -209,240 +190,3 @@ describe("Type Helpers", () => {
     });
   });
 });
-
-// ============================================================================
-// Capability Function Tests
-// ============================================================================
-
-describe("Capability Functions", () => {
-  describe("getCapabilityById", () => {
-    it("returns capability by ID", () => {
-      const capability = getCapabilityById(testCategories, "scale");
-      assert.ok(capability);
-      assert.strictEqual(capability.id, "scale");
-      assert.strictEqual(capability.name, "Scale");
-    });
-
-    it("returns undefined for unknown ID", () => {
-      const capability = getCapabilityById(testCategories, "unknown");
-      assert.strictEqual(capability, undefined);
-    });
-  });
-
-  describe("getCapabilityOrder", () => {
-    it("returns capabilities sorted by order", () => {
-      const ordered = getCapabilityOrder(testCategories);
-      assert.strictEqual(ordered[0], "scale");
-      assert.strictEqual(ordered[1], "ai");
-      assert.strictEqual(ordered[2], "people");
-    });
-
-    it("handles empty array", () => {
-      const ordered = getCapabilityOrder([]);
-      assert.strictEqual(ordered.length, 0);
-    });
-  });
-
-  describe("getCapabilityEmoji", () => {
-    it("returns emoji for capability", () => {
-      const emoji = getCapabilityEmoji(testCategories, "scale");
-      assert.strictEqual(emoji, "📐");
-    });
-
-    it("returns default for unknown capability", () => {
-      const emoji = getCapabilityEmoji(testCategories, "unknown");
-      assert.strictEqual(emoji, "💡");
-    });
-  });
-
-  describe("getCapabilityResponsibility", () => {
-    it("returns responsibility for capability and level", () => {
-      const responsibility = getCapabilityResponsibility(
-        testCategories,
-        "scale",
-        "working",
-      );
-      assert.strictEqual(responsibility, "Design scalable components");
-    });
-
-    it("returns undefined for unknown capability", () => {
-      const responsibility = getCapabilityResponsibility(
-        testCategories,
-        "unknown",
-        "working",
-      );
-      assert.strictEqual(responsibility, undefined);
-    });
-
-    it("returns undefined for unknown level", () => {
-      const responsibility = getCapabilityResponsibility(
-        testCategories,
-        "scale",
-        "mythical",
-      );
-      assert.strictEqual(responsibility, undefined);
-    });
-  });
-});
-
-describe("Standard emoji function", () => {
-  describe("getConceptEmoji", () => {
-    const testStandard = {
-      entityDefinitions: {
-        driver: { emojiIcon: "🎯" },
-        skill: { emojiIcon: "💼" },
-        behaviour: { emojiIcon: "🧠" },
-        discipline: { emojiIcon: "🔧" },
-        level: { emojiIcon: "📊" },
-        track: { emojiIcon: "🛤️" },
-      },
-    };
-
-    it("returns emoji for valid concept", () => {
-      assert.strictEqual(getConceptEmoji(testStandard, "driver"), "🎯");
-      assert.strictEqual(getConceptEmoji(testStandard, "skill"), "💼");
-      assert.strictEqual(getConceptEmoji(testStandard, "behaviour"), "🧠");
-      assert.strictEqual(getConceptEmoji(testStandard, "discipline"), "🔧");
-      assert.strictEqual(getConceptEmoji(testStandard, "level"), "📊");
-      assert.strictEqual(getConceptEmoji(testStandard, "track"), "🛤️");
-    });
-
-    it("returns default emoji for unknown concept", () => {
-      const emoji = getConceptEmoji(testStandard, "unknown");
-      assert.strictEqual(emoji, "💡");
-    });
-
-    it("returns default emoji when standard is null", () => {
-      const emoji = getConceptEmoji(null, "driver");
-      assert.strictEqual(emoji, "💡");
-    });
-
-    it("returns default emoji when concept has no emoji", () => {
-      const standard = { entityDefinitions: { driver: { name: "Drivers" } } };
-      const emoji = getConceptEmoji(standard, "driver");
-      assert.strictEqual(emoji, "💡");
-    });
-  });
-});
-
-describe("deriveResponsibilities", () => {
-  it("returns empty array when no capabilities provided", () => {
-    const skillMatrix = [
-      { skillId: "skill_a", capability: "scale", proficiency: "working" },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: [],
-    });
-    assert.strictEqual(result.length, 0);
-  });
-
-  it("excludes awareness-only capabilities", () => {
-    const skillMatrix = [
-      { skillId: "skill_a", capability: "scale", proficiency: "awareness" },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: testCategories,
-    });
-    assert.strictEqual(result.length, 0);
-  });
-
-  it("includes capabilities based on skill proficiency", () => {
-    // skill_a is in testDiscipline.coreSkills and has capability "scale"
-    const skillMatrix = [
-      { skillId: "skill_a", capability: "scale", proficiency: "working" },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: testCategories,
-    });
-    assert.strictEqual(result.length, 1);
-    assert.strictEqual(result[0].capability, "scale");
-    assert.strictEqual(result[0].proficiency, "working");
-  });
-
-  it("uses highest skill proficiency in each capability", () => {
-    const skillMatrix = [
-      { skillId: "skill_a", capability: "scale", proficiency: "working" },
-      {
-        skillId: "skill_extra",
-        capability: "scale",
-        proficiency: "practitioner",
-      },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: testCategories,
-    });
-    // Should use practitioner level for scale capability
-    const scaleResp = result.find((r) => r.capability === "scale");
-    assert.ok(scaleResp);
-    assert.strictEqual(scaleResp.proficiency, "practitioner");
-    assert.strictEqual(
-      scaleResp.responsibility,
-      "Lead architectural decisions",
-    );
-  });
-
-  it("includes responsibility from capability definition", () => {
-    const skillMatrix = [
-      { skillId: "skill_b", capability: "ai", proficiency: "working" },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: testCategories,
-    });
-    const aiResp = result.find((r) => r.capability === "ai");
-    assert.ok(aiResp);
-    assert.strictEqual(aiResp.responsibility, "Integrate AI capabilities");
-    assert.strictEqual(aiResp.proficiency, "working");
-  });
-
-  it("includes emoji from capability", () => {
-    const skillMatrix = [
-      { skillId: "skill_a", capability: "scale", proficiency: "working" },
-    ];
-    const result = deriveResponsibilities({
-      skillMatrix,
-      capabilities: testCategories,
-    });
-    assert.strictEqual(result[0].emojiIcon, "📐");
-  });
-});
-
-describe("deriveJob with capabilities", () => {
-  it("includes derived responsibilities when capabilities provided", () => {
-    const job = deriveJob({
-      discipline: testDiscipline,
-      level: testLevel,
-      track: testTrack,
-      skills: testSkills,
-      behaviours: testBehaviours,
-      capabilities: testCategories,
-    });
-
-    assert.ok(job);
-    assert.ok(job.derivedResponsibilities);
-    assert.ok(Array.isArray(job.derivedResponsibilities));
-    assert.ok(job.derivedResponsibilities.length > 0);
-  });
-
-  it("returns empty responsibilities when no capabilities provided", () => {
-    const job = deriveJob({
-      discipline: testDiscipline,
-      level: testLevel,
-      track: testTrack,
-      skills: testSkills,
-      behaviours: testBehaviours,
-    });
-
-    assert.ok(job);
-    assert.ok(Array.isArray(job.derivedResponsibilities));
-    assert.strictEqual(job.derivedResponsibilities.length, 0);
-  });
-});
-
-// ============================================================================
-// Validation Tests
-// ============================================================================
